requirements.txt
gradio>=4.0.0
tensorflow>=2.13.0
numpy
Pillow
